Wip 0.1.3.3

+++ src/duHast/Utilities/batch_processor_log_utils.py
@@ -51,7 +51,7 @@ Helper functions to process Revit BatchProcessor log files.
 
 import clr
 import System
-from System.IO import Path
+
 import glob
 import time
 import os
@@ -166,7 +166,7 @@ def write_session_id_marker_file(folder_path, session_id):
     return status
 
 
-def get_current_session_ids(folder_path):
+def get_current_session_ids(folder_path, delete_marker_files=True):
     """
     Returns file names of all text files in a given directory representing session Ids.
 
@@ -180,16 +180,19 @@ def get_current_session_ids(folder_path):
     """
     ids = []
     file_list = glob.glob(folder_path + "\\*" + ".txt")
-    # delete marker files
-    result_delete = True
-    for fd in file_list:
-        if debug_mode_ == False:
-            result_delete = result_delete & fileIO.file_delete(fd)
-    if not result_delete:
-        output("Failed to delete a marker file!")
+    # delete marker files if required
+    if(delete_marker_files == True):
+        for fd in file_list:
+            fileIO.file_delete(fd)
+        result_delete = True
+        for fd in file_list:
+            if debug_mode_ == False:
+                result_delete = result_delete & fileIO.file_delete(fd)
+        if not result_delete:
+            output("Failed to delete a marker file!")
     for f in file_list:
         ids.append(
-            adjust_session_id_file_name_back(Path.GetFileNameWithoutExtension(f))
+            adjust_session_id_file_name_back(fileIO.get_file_name_without_ext(f))
         )
     return ids
